+++ api/src/main/java/net/jqwik/api/PropertyDefaults.java
@@ -30,7 +30,7 @@ public @interface PropertyDefaults {
 
 	EdgeCasesMode edgeCases() default EdgeCasesMode.NOT_SET;
 
-	@API(status = EXPERIMENTAL, since = "1.4.0")
+	@API(status = MAINTAINED, since = "1.4.0")
 	FixedSeedMode whenFixedSeed() default FixedSeedMode.NOT_SET;
 
 	class PropertyDefaultsHook implements AroundPropertyHook {