+++ api/src/main/java/net/jqwik/api/constraints/NotBlank.java
@@ -17,6 +17,6 @@ import static org.apiguardian.api.API.Status.*;
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@API(status = EXPERIMENTAL, since = "1.5.0")
+@API(status = MAINTAINED, since = "1.5.0")
 public @interface NotBlank {
 }