+++ api/src/main/java/net/jqwik/api/domains/DomainContextBase.java
@@ -37,7 +37,7 @@ import static org.apiguardian.api.API.Status.*;
  * @see DomainContext
  * @see Provide
  */
-@API(status = EXPERIMENTAL, since = "1.5.2")
+@API(status = MAINTAINED, since = "1.5.2")
 @NonNullApi
 public abstract class DomainContextBase implements DomainContext {