+++ web/src/main/java/net/jqwik/web/api/Email.java
@@ -21,7 +21,7 @@ import static org.apiguardian.api.API.Status.*;
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.PARAMETER, ElementType.TYPE_USE })
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@API(status = EXPERIMENTAL, since = "1.4.0")
+@API(status = MAINTAINED, since = "1.4.0")
 public @interface Email {
 
 	/**